[Init/#1] hilt, buildConfig, move
