+++ app/src/main/java/com/swyp/firsttodo/Application.kt
@@ -1,0 +1,19 @@
+package com.swyp.firsttodo
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class Application : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        setTimber()
+    }
+
+    private fun setTimber() {
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
