+++ app/src/main/java/com/swyp/firsttodo/core/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.ui.theme
+package com.swyp.firsttodo.core.designsystem.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
