+++ app/src/main/java/com/swyp/firsttodo/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.ui.theme
+package com.swyp.firsttodo.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
