+++ app/src/main/java/com/swyp/firsttodo/presentation/main/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.swyp.firsttodo.presentation.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.swyp.firsttodo.core.designsystem.theme.FirstTodoTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            FirstTodoTheme {
+                MainScreen()
+            }
+        }
+    }
+}
