+++ app/src/main/java/com/swyp/firsttodo/presentation/main/MainScreen.kt
@@ -1,0 +1,19 @@
+package com.swyp.firsttodo.presentation.main
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun MainScreen() {
+    Scaffold(
+        modifier = Modifier.fillMaxSize(),
+    ) { innerPadding ->
+        Box(
+            modifier = Modifier.padding(innerPadding),
+        )
+    }
+}
